# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Future Me API",
    "description": "Future Me is a time capsule letter platform that allows users to send letters to their future self or loved ones, scheduled for delivery on a future date. This API provides endpoints for user authentication, letter creation and management, draft system, scheduled delivery, notifications, and media attachments.",
    "version": "1.0.0",
    "contact": {
      "name": "Future Me Support",
      "email": "support@futureme.com"
    }
  },
  "servers": [
    {
      "url": "https://backend.hyfuture.app/api/v1",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:4001/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token. Get it from the login or verify-auth-code endpoints."
      }
    },
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        }
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "Jane Doe",
            "description": "Optional user name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "newemail@example.com",
            "description": "Optional email address. If changed, email verification will be reset."
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": ["currentPassword", "newPassword", "confirmPassword"],
        "properties": {
          "currentPassword": {
            "type": "string",
            "format": "password",
            "example": "Oldpass123!",
            "description": "Current password"
          },
          "newPassword": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "maxLength": 100,
            "example": "Newpass123!",
            "description": "New password"
          },
          "confirmPassword": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "maxLength": 100,
            "example": "Newpass123!",
            "description": "Confirm new password (must match newPassword)"
          }
        }
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "ce285529-0508-4112-bd64-56d4156af2d6"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T10:00:00.000Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Get Profile",
        "description": "Retrieve the authenticated user's profile information. Password is never included in the response.",
        "operationId": "getProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                },
                "example": {
                  "id": "ce285529-0508-4112-bd64-56d4156af2d6",
                  "name": "John Doe",
                  "email": "user@example.com",
                  "isEmailVerified": true,
                  "createdAt": "2024-01-01T10:00:00.000Z",
                  "updatedAt": "2024-01-01T10:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Profile"],
        "summary": "Update Profile",
        "description": "Update the authenticated user's profile information. Only provided fields will be updated. If email is changed, email verification will be reset.",
        "operationId": "updateProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              },
              "example": {
                "name": "Jane Doe",
                "email": "newemail@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                },
                "example": {
                  "id": "ce285529-0508-4112-bd64-56d4156af2d6",
                  "name": "Jane Doe",
                  "email": "newemail@example.com",
                  "isEmailVerified": false,
                  "createdAt": "2024-01-01T10:00:00.000Z",
                  "updatedAt": "2024-01-01T11:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid email or email already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/profile/password": {
      "patch": {
        "tags": ["Profile"],
        "summary": "Change Password",
        "description": "Change the authenticated user's password. Requires current password verification.",
        "operationId": "changePassword",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              },
              "example": {
                "currentPassword": "Oldpass123!",
                "newPassword": "Newpass123!",
                "confirmPassword": "Newpass123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "Password changed successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Password mismatch or new password same as current",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "New password and confirm password do not match",
                  "error": "Bad Request"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Current password is incorrect or missing/invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Current password is incorrect",
                  "error": "Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
  }
}